requirements.txt
fastapi
uvicorn
pydantic
python-multipart
langgraph
torch
torchvision
pandas
pillow
